Split Encoder and Decoder to separate files
